+++ packages/hspec/src/Conferer/FromConfig/Hspec.hs
@@ -33,8 +33,12 @@ instance FromConfig Hspec.Formatter where
     fetchFromConfigWith $
     (\case
       "silent" -> Just Hspec.silent
+#if MIN_VERSION_hspec_core(2,7,10)
+      "checks" -> Just Hspec.checks
+#endif
       "specdoc" -> Just Hspec.specdoc
       "progress" -> Just Hspec.progress
+      "failed-examples" -> Just Hspec.failed_examples
       "failed_examples" -> Just Hspec.failed_examples
       _ -> Nothing
     ) . toLower
@@ -75,6 +79,9 @@ desconstructHspecConfigToDefaults Hspec.Config{..} =
   , ("focusedOnly", toDyn configFocusedOnly)
   , ("failOnFocused", toDyn configFailOnFocused)
 #endif
+#if MIN_VERSION_hspec_core(2,7,3)
+  , ("randomize", toDyn configRandomize)
+#endif
   ]
 
 instance FromConfig Hspec.Config where
@@ -114,5 +121,8 @@ instance FromConfig Hspec.Config where
 #if MIN_VERSION_hspec_core(2,7,0)
     configFocusedOnly <- fetchFromConfig (key /. "focusedOnly") config
     configFailOnFocused <- fetchFromConfig (key /. "failOnFocused") config
+#endif
+#if MIN_VERSION_hspec_core(2,7,3)
+    configRandomize <- fetchFromConfig (key /. "randomize") config
 #endif
     pure Hspec.Config{..}